finish moving over new grid traits
