+++ traits/src/cell.rs
@@ -1,4 +1,5 @@
 use crate::element::FiniteElement;
+use crate::types::ReferenceCellType;
 use std::fmt;
 
 #[derive(Debug)]
@@ -11,19 +12,6 @@ impl fmt::Display for InvalidConnectivity {
 }
 
 impl std::error::Error for InvalidConnectivity {}
-
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
-#[repr(u8)]
-pub enum ReferenceCellType {
-    Point = 0,
-    Interval = 1,
-    Triangle = 2,
-    Quadrilateral = 3,
-    Tetrahedron = 4,
-    Hexahedron = 5,
-    Prism = 6,
-    Pyramid = 7,
-}
 
 /// A 0- to 3- dimensional reference cell
 pub trait ReferenceCell {
