+++ traits/src/grid.rs
@@ -1,17 +1,17 @@
 //! Trait definitions for a grid
 
-pub mod builder;
-pub mod cell;
-pub mod geometry;
-pub mod grid;
-pub mod point;
-pub mod reference_map;
-pub mod topology;
+mod builder;
+mod cell;
+mod geometry;
+mod grid_type;
+mod point;
+mod reference_map;
+mod topology;
 
 pub use builder::*;
 pub use cell::*;
 pub use geometry::*;
-pub use grid::*;
+pub use grid_type::*;
 pub use point::*;
 pub use reference_map::*;
 pub use topology::*;
